chore: reorg types/payment-types
